+++ packages/playground/src/index.tsx
@@ -20,17 +20,15 @@
 
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { GlobalStyle } from '@looker/components'
-import { theme } from '@looker/design-tokens'
-import { ThemeProvider } from 'styled-components'
+import { ComponentsProvider, GlobalStyle } from '@looker/components'
 import { ActionListDemo } from './ActionList/ActionListDemo'
 
 const App: React.FC = () => {
   return (
-    <ThemeProvider theme={theme}>
+    <ComponentsProvider>
       <GlobalStyle />
       <ActionListDemo />
-    </ThemeProvider>
+    </ComponentsProvider>
   )
 }