+++ packages/www/src/MDX/Pre/allComponents.ts
@@ -46,6 +46,7 @@ import {
   Code,
   CodeBlock,
   ColorWheel,
+  ComponentsProvider,
   Confirm,
   ConfirmLayout,
   DateFormat,
@@ -97,6 +98,7 @@ import {
   Paragraph,
   Popover,
   PopoverContent,
+  Prompt,
   Radio,
   Select,
   SelectMulti,
@@ -128,7 +130,7 @@ import {
 } from '@looker/components'
 import { palette } from '@looker/design-tokens'
 
-import styled, { ThemeProvider } from 'styled-components'
+import styled from 'styled-components'
 import isEqual from 'lodash/isEqual'
 import { useState } from 'react'
 
@@ -136,7 +138,6 @@ export const allComponents = {
   useState,
 
   styled,
-  ThemeProvider,
 
   isEqual,
 
@@ -162,6 +163,7 @@ export const allComponents = {
   Code,
   CodeBlock,
   ColorWheel,
+  ComponentsProvider,
   Confirm,
   ConfirmLayout,
   DateFormat,
@@ -213,6 +215,7 @@ export const allComponents = {
   Paragraph,
   Popover,
   PopoverContent,
+  Prompt,
   Radio,
   Select,
   SelectMulti,